Workflow Automation/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyYAML
google-generativeai
fastapi
uvicorn
sqlalchemy